Move hopper sensor to IndexerIOHardware; gate shot bindings with isHopperFull
Root cause: CANrange in IntakeIOHardware shared the same hardware init and
refresh cycle as the slide TalonFX, causing slide control interference.

Fix:
- Fully revert IntakeIO, IntakeIOHardware, IntakeSubsystem (no sensor there)
- Add hopper CANrange (ID 41) to IndexerIOHardware alongside existing chute
  sensor — same config pattern, added to BaseStatusSignal.refreshAll()
- Add hopperDistanceMeters + hopperFull to IndexerIOInputs
- Add isHopperFull() to IndexerSubsystem
- RobotContainer: apply .unless(indexer::isHopperFull) only on shot deadlines
  (driver rightTrigger/rightBumper, operator a/b/x/y); manual rightBumper
  compression is unchanged so it always runs

fuelCompression() itself is completely unchanged.
