+++ src/main/java/frc/robot/subsystems/intake/IntakeSubsystem.java
@@ -2,7 +2,6 @@ package frc.robot.subsystems.intake;
 
 import java.util.function.BooleanSupplier;
 
-import edu.wpi.first.networktables.BooleanPublisher;
 import edu.wpi.first.networktables.DoublePublisher;
 import edu.wpi.first.networktables.NetworkTable;
 import edu.wpi.first.networktables.NetworkTableInstance;
@@ -48,8 +47,6 @@ public class IntakeSubsystem extends SubsystemBase {
     private final NetworkTable intakeTable;
     private final StringPublisher intakeStatePublisher;
     private final DoublePublisher slidePositionPublisher;
-    private final DoublePublisher hopperDistancePublisher;
-    private final BooleanPublisher hopperFullPublisher;
 
     // =====================================================================
     // Constructor
@@ -59,10 +56,8 @@ public class IntakeSubsystem extends SubsystemBase {
 
         NetworkTableInstance inst = NetworkTableInstance.getDefault();
         intakeTable = inst.getTable("Intake");
-        intakeStatePublisher     = intakeTable.getStringTopic("State").publish();
-        slidePositionPublisher   = intakeTable.getDoubleTopic("SlidePosition").publish();
-        hopperDistancePublisher  = intakeTable.getDoubleTopic("HopperDistance").publish();
-        hopperFullPublisher      = intakeTable.getBooleanTopic("HopperFull").publish();
+        intakeStatePublisher = intakeTable.getStringTopic("State").publish();
+        slidePositionPublisher = intakeTable.getDoubleTopic("SlidePosition").publish();
     }
 
     // =====================================================================
@@ -79,8 +74,6 @@ public class IntakeSubsystem extends SubsystemBase {
     private void publishTelemetry() {
         intakeStatePublisher.set(getIntakeState());
         slidePositionPublisher.set(inputs.slidePositionRotations);
-        hopperDistancePublisher.set(inputs.hopperDistanceMeters);
-        hopperFullPublisher.set(inputs.hopperFull);
     }
 
     // =====================================================================
@@ -116,11 +109,6 @@ public class IntakeSubsystem extends SubsystemBase {
     /** True when the slide is safely past the home position for roller operation. */
     public boolean isSlidePastHome() {
         return inputs.slidePositionRotations > Constants.Intake.SLIDE_ROLLER_SAFE_POS;
-    }
-
-    /** True when the hopper top sensor detects a ball within HOPPER_FULL_DISTANCE_METERS. */
-    public boolean isHopperFull() {
-        return inputs.hopperFull;
     }
 
     // Position accessor for commands and calculations.
@@ -414,7 +402,6 @@ public class IntakeSubsystem extends SubsystemBase {
         return compressFuel(
                 Constants.Intake.SLIDE_FUEL_COMPRESSION_WAIT_SECONDS,
                 Constants.Intake.SLIDE_FUEL_COMPRESSION_DURATION_SECONDS)
-                .unless(this::isHopperFull)
                 .withName("FuelCompression");
     }
